Reference

Reference

Creating a hypergraph

SimpleHypergraphs.Hypergraph — Type.

Hypergraph{T} <: AbstractMatrix{Union{T, Nothing}}

A hypergraph storing information about vertices and hyperedges.

Constructors

Hypergraph{T}(n,k) where {T<:Real}
Hypergraph{T,V}(n, k) where {T<:Real, V}
Hypergraph{T,V,E}(n, k) where {T<:Real, V, E}

Construct a hypergraph with a given number of vertices and hyperedges. Optionally, values of type V can be stored at vertices and values of type E can be stored at hyperedges.

Hypergraph(m::AbstractMatrix{Union{T, Nothing}}) where {T<:Real}
Hypergraph{V}(m::AbstractMatrix{Union{T, Nothing}}) where {T<:Real, V}
Hypergraph{V, E}(m::AbstractMatrix{Union{T, Nothing}}) where {T<:Real, V, E}

Construct a hypergraph using its matrix representation. In the matrix representation rows are vertices and columns are hyperedges. Optionally, values of type V can be stored at vertices and values of type E can be stored at hyperedges.

Hypergraph(g::LightGraphs.Graph)

Constructs a hypergraph of degree 2 by making a deep copy of LightGraphs.Graph

Arguments

T : type of weight values stored in the hypergraph

V : type of values stored in the vertices of the hypergraph

E : type of values stored in the edges of the hypergraph

n : number of vertices

k : number of hyperedges

m : a matrix representation rows are vertices and columns are hyperedges

source

Manipulating vertices and hyperedges

SimpleHypergraphs.add_hyperedge! — Method.

add_hyperedge!(h::Hypergraph{T, V, E}; vertices::Dict{Int,T} = Dict{Int,T}(),
               hyperedge_meta::Union{E,Nothing}=nothing) where {T <: Real, V, E}

Adds a hyperedge to a given hypergraph h. Optionally, existing vertices can be added to the created hyperedge. The paramater vertices represents a dictionary of vertex identifiers and values stored at the hyperedges. Additionally, a value can be stored with the hyperedge using the hyperedge_meta keyword parameter.

source

SimpleHypergraphs.add_vertex! — Method.

add_vertex!(h::Hypergraph{T, V, E}; hyperedges::Dict{Int,T} = Dict{Int,T}(),
            vertex_meta::Union{V,Nothing} = nothing) where {T <: Real, V, E}

Adds a vertex to a given hypergraph h. Optionally, the vertex can be added to existing hyperedges. The hyperedges parameter presents a dictionary of hyperedge identifiers and values stored at the hyperedges. Additionally, a value can be stored with the vertex using the vertex_meta keyword parameter.

source

SimpleHypergraphs.set_vertex_meta! — Method.

set_vertex_meta!(h::Hypergraph{T, V, E}, new_value::Union{V,Nothing}, id::Int)
    where {T <: Real, V, E}

Sets a new meta value new_value for the vertex id in the hypegraph h.

source

SimpleHypergraphs.get_vertex_meta — Method.

get_vertex_meta(h::Hypergraph{T, V, E}, id::Int) where {T <: Real, V, E}

Returns a meta value stored at the vertex id in the hypergraph h.

source

SimpleHypergraphs.set_hyperedge_meta! — Method.

set_hyperedge_meta!(h::Hypergraph{T, V, E}, new_value::Union{E,Nothing}, id::Int)
    where {T <: Real, V, E}

Sets a new meta value new_value for the hyperedge id in the hypegraph h.

source

SimpleHypergraphs.get_hyperedge_meta — Method.

get_hyperedge_meta(h::Hypergraph{T, V, E}, id::Int) where {T <: Real, V, E}

Returns a meta value stored at the hyperedge id in the hypergraph h.

source

SimpleHypergraphs.remove_vertex! — Method.

remove_vertex!(h::Hypergraph, v::Int)

Removes the vertex v from a given hypergraph h. Note that running this function will cause reordering of vertices in the hypergraph: the vertex v will replaced by the last vertex of the hypergraph and the list of vertices will be shrunk.

source

Hypergraph array getters and setters

Normally you work with a hypergraph via array setters, for example the code below craete an Hypergraph and add vertex one to hyperedges 2 and 3 with weight 5:

h = Hypergraph{Int64}(2,3);
h[1, 2:3] .= 5;  
h

# output

2×3 Hypergraph{Int64,Nothing,Nothing}:
 nothing  5         5
 nothing   nothing   nothing

Base.getindex — Method.

Base.getindex(h::Hypergraph, idx::Vararg{Int,2})

Returns a value for a given vertex-hyperedge pair idx for a hypergraph h. If a vertex does not belong to a hyperedge nothing is returned.

source

Base.setindex! — Method.

Base.setindex!(h::Hypergraph, ::Nothing, idx::Vararg{Int,2})

Removes a vertex from a given hyperedge for a hypergraph h and a given vertex-hyperedge pair idx. Note that trying to remove a vertex from a hyperedge when it is not present will not throw an error.

source

Base.setindex! — Method.

Base.setindex!(h::Hypergraph, v::Real, idx::Vararg{Int,2})

Adds a vertex to a hyperedge (represented by indices idx) and assigns value v to be stored with that assignment.

source

Hypergraph representation as LightGraphs' simple graphs

The goal of those methods is to provide a way to manipulate a hypergraph using the methods from the LightGraphs.jl library. This has been achieved by providing types that are subtypes of the LightGraphs.SimpleGraphs.AbstractSimpleGraph{Int} type along with appropiate methods.

SimpleHypergraphs.BipartiteView — Type.

BipartiteView{T<:Real} <: LightGraphs.SimpleGraphs.AbstractSimpleGraph{Int}

Represents a bipartite view of a hypergraph h. Note this is a view - changes to the original hypergraph will be automatically reflected in the view.

Constructors

BipartiteView(::Hypergraph)

The bipartite view of a hypergraph is suitable for processing with the LightGraphs.jl package. Several LightGraphs methods are provided for the compability.

source

SimpleHypergraphs.shortest_path — Method.

shortest_path(b::BipartiteView,source::Int, target::Int)

Finds a single shortest path in a graph b between vertices source and target. Note that if several paths of the same length exist, only one will be returned.

source

SimpleHypergraphs.TwoSectionView — Type.

TwoSectionView{T<:Real} <: LightGraphs.SimpleGraphs.AbstractSimpleGraph{Int64}

Represents a 2-section view of a hypergraph h. Note this is a view - changes to the original hypergraph will be automatically reflected in the view.

Constructors

TwoSectionView(::Hypergraph)

The 2-section view of a hypergraph is suitable for processing with the LightGraphs.jl package. Several LightGraphs methods are provided for the compability.

source

SimpleHypergraphs.shortest_path — Method.

shortest_path(t::TwoSectionView,source::Int, target::Int)

Finds a single shortest path in a graph b between vertices source and target. Note that if several paths of the same length exist, only one will be returned.

source

Hypergraph info

Base.size — Method.

Base.size(h::Hypergraph)

Returns the size of Hypergraph h. The result is a tuple of the number of vertices and the number of hyperedges

source

SimpleHypergraphs.nhv — Method.

nhv(h::Hypergraph{T, V, E}) where {T <: Real, V, E}

Return the number of vertices in the hypergraph h.

source

SimpleHypergraphs.nhe — Method.

nhe(h::Hypergraph{T, V, E}) where {T <: Real, V, E}

Return the number of hyperedges in the hypergraph h.

source

SimpleHypergraphs.getvertices — Method.

getvertices(h::Hypergraph, he_id::Int)

Returns vertices from a hypergraph a for a given hyperedge he_id.

source

SimpleHypergraphs.gethyperedges — Method.

gethyperedges(h::Hypergraph, v_id::Int)

Returns hyperedges for a given vertex v_id in a hypergraph h.

source

SimpleHypergraphs.modularity — Method.

modularity(h::Hypergraph, partition::Vector{Set{Int}}, ha::HypergraphAggs=HypergraphAggs(h))

Calculates the strict modularity of a hypergraph h for a given partition using the precomputed aggregates ha.

source

SimpleHypergraphs.HypergraphAggs — Type.

HypergraphAggs(h::Hypergraph)

Precomputes vertex and edge basic stats for a hypergraph. The stats are being used for efficiency reasons by community search algorithms.

source

SimpleHypergraphs.randompartition — Method.

randompartition(N::Int, n::Int)::Vector{Set{Int}}

Generates a random partition for graph having N vertices into n subsets.

source

SimpleHypergraphs.randompartition — Method.

randompartition(h::Hypergraph, n::Int)::Vector{Set{Int}}

Generates a random partition for vertices of a hypergraph h into n subsets.

source

SimpleHypergraphs.AbstractCommunityFinder — Type.

The base type for all algorithms representing various community search patterns.

source

SimpleHypergraphs.CFModularityRandom — Type.

CFModularityRandom(n::Int, reps::Int) <: AbstractCommunityFinder

Represents a random search over the hypergraph h that finds a partition into n communities (subsets) having the maximum modularity value. During the search reps random n-partitions will be evaluated. If there are many partitions having the same value the first one that was randomly found will be returned.

source

SimpleHypergraphs.findcommunities — Method.

findcommunities(h::Hypergraph, method::CFModularityRandom)

Makes a random search over the hypergraph h and finds a partition into method.n communities (subsets) having the maximum modularity value. During the search method.reps random n-partitions will be evaluated. If there are many partitions having the same value the first one that was randomly found will be returned.

Returns a NamedTuple where the field bp contains partition and the field bm contains the modularity value for that partition.

source

I/O

SimpleHypergraphs.hg_save — Function.

hg_save(io::IO, h::Hypergraph)

Saves a hypergraph h to an output stream io.

source

hg_save(fname::AbstractString, h::Hypergraph)

Saves a hypergraph h to a file fname.

source

SimpleHypergraphs.hg_load — Function.

hg_load(fname::AbstractString, T::Type{<:Real})

Loads a hypergraph from a stream io. The second argument T represents type of data in the hypegraph.

Skips an initial comment.

source

hg_load(fname::AbstractString, T::Type{<:Real})

Loads a hypergraph from a file fname. The second argument T represents type of data in the hypegraph

source