Reference

Reference

Types

SimpleHypergraphs.Hypergraph — Type.

Hypergraph{T} <: AbstractMatrix{Union{T, Nothing}}

A hypergraph storing information about vertices and hyperedges.

Constructors

Hypergraph{T}(n,k) where {T<:Real}

Construct a hypergraph with a given number of vertices and hyperedges.

Hypergraph(m::AbstractMatrix{T}) where {T<:Real}

Construct a hypergraph using its matrix representation. In the matrix representation rows are vertices and columns are hyperedges.

Arguments

T : type of values stored in the hypergraph

n : number of vertices

k : number of hyperedges

m : a matrix representation rows are vertices and columns are hyperedges

source

Hypergraph array getters and setters

Normally you work with a hypergraph via array setters, for example the code below craete an Hypergraph and add vertex one to hyperedges 2 and 3 with weight 5:

h = Hypergraph{Int64}(2,3);
h[1, 2:3] .= 5;  
h

# output

2×3 Hypergraph{Int64}:
 nothing  5         5
 nothing   nothing   nothing

Base.getindex — Method.

Base.getindex(h::Hypergraph, idx::Vararg{Int,2})

Returns a value for a given vertex-hyperedge pair or nothing if a vertex does not belong to a hyperedge.

Arguments

h : a hypergraph

idx : an index where the first element is vertex and the second is a hyperedge

source

Base.setindex! — Method.

Base.setindex!(h::Hypergraph, ::Nothing, idx::Vararg{Int,2})

Removes a vertex from a given hyperedge

Arguments

h : a hypergraph

idx : an index where the first element is vertex and the second is a hyperedge

source

Base.setindex! — Method.

Base.setindex!(h::Hypergraph, v::Real, idx::Vararg{Int,2})

Adds a vertex to a hyperedge and assigns value to be stored with that assignment.

Arguments

h : a hypergraph

v : a value to be stored with vertex-hyperedge assignment

idx : an index where the first element is vertex and the second is a hyperedge

source

Hypergraph info

Base.size — Method.

Base.size(h::Hypergraph)

Returns the size of Hypergraph m. The result is a tuple of the number of vertices and the number of hyperedges

Arguments

h : a hypergraph

source

SimpleHypergraphs.getvertices — Method.

getvertices(h::Hypergraph, he_id::Int)

Returns vertices for a given hyperedge

Arguments

h : a hypergraph

he_id : an identifier of a hyperedge

source

SimpleHypergraphs.gethyperedges — Method.

gethyperedges(h::Hypergraph, v_id::Int)

Returns hyperedges for a given vertex

Arguments

h : a hypergraph

v_id : an identifier of a vertex

source

I/O

SimpleHypergraphs.hg_save — Function.

hg_save(io::IO, h::Hypergraph)

Saves a hypergraph to a stream.

Arguments

io : the output stream to be written to

h : a hypergraph

source

hg_save(fname::AbstractString, h::Hypergraph)

Saves a hypergraph to a file.

Arguments

fname : a file name to save to

h : a hypergraph

source

SimpleHypergraphs.hg_load — Function.

hg_load(fname::AbstractString, T::Type{<:Real})

Loads a hypergraph from a stream.

Arguments

io : an input stream from which to load the data

T : type of data in the hypegraph

source

hg_load(fname::AbstractString, T::Type{<:Real})

Loads a hypergraph from a file.

Arguments

fname : a file from which to load the data

T : type of data in the hypegraph

source