Reference

Reference

Types

SimpleHypergraphs.Hypergraph — Type.

Hypergraph{T} <: AbstractMatrix{Union{T, Nothing}}

A hypergraph storing information about vertices and hyperedges.

Constructors

Hypergraph{T}(n,k) where {T<:Real}

Construct a hypergraph with a given number of vertices and hyperedges.

Hypergraph(m::AbstractMatrix{T}) where {T<:Real}

Construct a hypergraph using its matrix representation. In the matrix representation rows are vertices and columns are hyperedges.

Arguments

T : type of values stored in the hypergraph

n : number of vertices

k : number of hyperedges

m : a matrix representation rows are vertices and columns are hyperedges

source

Manipulating vertices and hyperedges

SimpleHypergraphs.add_hyperedge! — Method.

add_hyperedge!(h::Hypergraph{T};vertices::Dict{Int,T} = Dict{Int,T}()) where T <: Real

Adds a hyperedge to a given hypergraph h. Optionally, existing vertices can be added to the created hyperedge. The paramater vertices represents a dictionary of vertex identifiers and values stored at the hyperedges

source

SimpleHypergraphs.add_vertex! — Method.

add_vertex!(h::Hypergraph{T};hyperedges::Dict{Int,T} = Dict{Int,T}()) where T <: Real

Adds a vertex to a given hypergraph h. Optionally, the vertex can be added to existing hyperedges. The hyperedges parameter presents a dictionary of hyperedge identifiers and values stored at the hyperedges

source

Hypergraph array getters and setters

Normally you work with a hypergraph via array setters, for example the code below craete an Hypergraph and add vertex one to hyperedges 2 and 3 with weight 5:

h = Hypergraph{Int64}(2,3);
h[1, 2:3] .= 5;  
h

# output

2×3 Hypergraph{Int64}:
 nothing  5         5
 nothing   nothing   nothing

Base.getindex — Method.

Base.getindex(h::Hypergraph, idx::Vararg{Int,2})

Returns a value for a given vertex-hyperedge pair idx for a hypergraph h. If a vertex does not belong to a hyperedge nothing is returned.

source

Base.setindex! — Method.

Base.setindex!(h::Hypergraph, ::Nothing, idx::Vararg{Int,2})

Removes a vertex from a given hyperedge for a hypergraph h and a given vertex-hyperedge pair idx.

source

Base.setindex! — Method.

Base.setindex!(h::Hypergraph, v::Real, idx::Vararg{Int,2})

Adds a vertex to a hyperedge (represented by indices idx) and assigns value v to be stored with that assignment.

source

Hypergraph info

Base.size — Method.

Base.size(h::Hypergraph)

Returns the size of Hypergraph h. The result is a tuple of the number of vertices and the number of hyperedges

source

SimpleHypergraphs.getvertices — Method.

getvertices(h::Hypergraph, he_id::Int)

Returns vertices from a hypergraph a for a given hyperedge he_id.

source

SimpleHypergraphs.gethyperedges — Method.

gethyperedges(h::Hypergraph, v_id::Int)

Returns hyperedges for a given vertex v_id in a hypergraph h.

source

SimpleHypergraphs.BipartiteView — Type.

BipartiteView{T<:Real} <: AbstractGraph{Int64}

Represents a bipartite view of a hypergraph h. Note this is a view - changes to the original hypergraph will be automatically reflected in the view.

Constructors

BipartiteView(::Hypergraph)

The bipartite view of a hypergraph is suitable for processing with the LightGraphs.jl package. Several LightGraphs methods are provided for the compability.

source

SimpleHypergraphs.shortest_path — Method.

shortest_path(b::BipartiteView,source::Int, target::Int)

Finds a single shortest path in a graph b between vertices source and target. Note that if several paths of the same length exist, only one will be returned.

source

SimpleHypergraphs.TwoSectionView — Type.

TwoSectionView{T<:Real} <: AbstractGraph{Int64}

Represents a 2-section view of a hypergraph h. Note this is a view - changes to the original hypergraph will be automatically reflected in the view.

Constructors

TwoSectionView(::Hypergraph)

The 2-section view of a hypergraph is suitable for processing with the LightGraphs.jl package. Several LightGraphs methods are provided for the compability.

source

SimpleHypergraphs.shortest_path — Method.

shortest_path(t::TwoSectionView,source::Int, target::Int)

Finds a single shortest path in a graph b between vertices source and target. Note that if several paths of the same length exist, only one will be returned.

source

I/O

SimpleHypergraphs.hg_save — Function.

hg_save(io::IO, h::Hypergraph)

Saves a hypergraph h to an output stream io.

source

hg_save(fname::AbstractString, h::Hypergraph)

Saves a hypergraph h to a file fname.

source

SimpleHypergraphs.hg_load — Function.

hg_load(fname::AbstractString, T::Type{<:Real})

Loads a hypergraph from a stream io. The second argument T represents type of data in the hypegraph

source

hg_load(fname::AbstractString, T::Type{<:Real})

Loads a hypergraph from a file fname. The second argument T represents type of data in the hypegraph

source